requirements.txt
click==7.1.2
dateparser==1.0.0
Flask==1.1.2
Flask-FontAwesome==0.1.5
Flask-SQLAlchemy==2.5.1
Flask-Testing==0.8.1
greenlet==1.0.0
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
marshmallow==3.11.1
marshmallow-sqlalchemy==0.24.2
pyserial==3.5
python-dateutil==2.8.1
pytz==2021.1
regex==2021.4.4
six==1.15.0
SQLAlchemy==1.4.11
tzlocal==2.1
Werkzeug==1.0.1
